修改原POP弹窗导致出现的问题。修改为Dialog处理

+++ app/src/main/java/com/dl/playfun/entity/GameCoinWalletEntity.java
@@ -32,7 +32,7 @@ public class GameCoinWalletEntity extends BaseObservable {
 //    @SerializedName("realname")
     private String currencyName;
 //    @SerializedName("account_number")
-    private int totalAppCoins;
+    private long totalAppCoins;
 
     @Bindable
     public int getTotalCoins() {
@@ -63,11 +63,11 @@ public class GameCoinWalletEntity extends BaseObservable {
     }
 
     @Bindable
-    public int getTotalAppCoins() {
+    public long getTotalAppCoins() {
         return totalAppCoins;
     }
 
-    public void setTotalGameCoins(int totalGameCoins) {
+    public void setTotalGameCoins(long totalGameCoins) {
         this.totalAppCoins = totalGameCoins;
     }
 }
